Classpath+classloader related classes moved to cp package
