+++ appserver/extras/embedded/common/bootstrap/src/main/java/org/glassfish/uberjar/activator/UberJarGlassFishActivator.java
@@ -18,7 +18,7 @@
 package org.glassfish.uberjar.activator;
 
 import com.sun.enterprise.glassfish.bootstrap.cfg.BootstrapKeys;
-import com.sun.enterprise.glassfish.bootstrap.osgi.impl.OsgiPlatform;
+import com.sun.enterprise.glassfish.bootstrap.cfg.OsgiPlatform;
 
 import java.security.AccessController;
 import java.security.PrivilegedAction;
