+++ nucleus/core/bootstrap/src/main/java/com/sun/enterprise/glassfish/bootstrap/cfg/StartupContextCfg.java
@@ -16,8 +16,6 @@
 
 package com.sun.enterprise.glassfish.bootstrap.cfg;
 
-import com.sun.enterprise.glassfish.bootstrap.osgi.impl.OsgiPlatform;
-
 import java.io.File;
 import java.nio.file.Path;
 import java.util.Properties;
