+++ nucleus/core/bootstrap/src/main/java/com/sun/enterprise/glassfish/bootstrap/cp/KnopflerfishAdapter.java
@@ -14,21 +14,26 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package com.sun.enterprise.glassfish.bootstrap.osgi.impl;
+package com.sun.enterprise.glassfish.bootstrap.cp;
 
-import com.sun.enterprise.glassfish.bootstrap.cp.ClassPathBuilder;
+import com.sun.enterprise.glassfish.bootstrap.cfg.StartupContextCfg;
 
-/**
- * Non-OSGI platform adapter.
- * Does not have any config file, does not record OSGI framework jar files.
- */
-public class EmbeddedAdapter implements OsgiPlatformAdapter {
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Path;
 
-    /**
-     * Noop.
-     */
+public class KnopflerfishAdapter implements OsgiPlatformAdapter {
+
+    private static final String KF_HOME = "KNOPFLERFISH_HOME";
+    private final File kfHome;
+
+    public KnopflerfishAdapter(StartupContextCfg cfg) {
+        this.kfHome = cfg.getOsgiHome(KF_HOME, KF_HOME, Path.of("osgi", "knopflerfish.org", "osgi"));
+    }
+
+
     @Override
-    public void addFrameworkJars(ClassPathBuilder cpb) {
-        // nothing to do
+    public void addFrameworkJars(ClassPathBuilder cpb) throws IOException {
+        cpb.addJar(new File(kfHome, "framework.jar"));
     }
 }
