+++ nucleus/core/bootstrap/src/main/java/com/sun/enterprise/glassfish/bootstrap/osgi/EmbeddedOSGiGlassFishRuntimeBuilder.java
@@ -18,7 +18,7 @@
 package com.sun.enterprise.glassfish.bootstrap.osgi;
 
 import com.sun.enterprise.glassfish.bootstrap.cfg.BootstrapKeys;
-import com.sun.enterprise.glassfish.bootstrap.osgi.impl.OsgiPlatform;
+import com.sun.enterprise.glassfish.bootstrap.cfg.OsgiPlatform;
 
 import java.util.Hashtable;
